Move service tests to __tests__ folder (#37)
