--- src/services/auth/index.ts
@@ -1,1 +1,0 @@
-export { default } from './auth';
